Initial Branch Commit
Remove blank line from package
Fix test
Fiz animation behavior when animation is in a view that has reuse (e.g. a List()) as this can cause the animation to leak and pin CPU usage at 100%. The fix is to set the animation to a zero length animation when the view disappears.